pull up the rev 1.34 to rev 1.35 change from the head:
Remove an lval typecast.  Not only does this get gerbv to compile with
gcc4 and other non-gcc compilers, it was no longer needed as the types
involved were all doubles anyway.

+++ src/drill.c
@@ -217,8 +217,8 @@ parse_drillfile(gerb_file_t *fd)
 	    /* KLUDGE. This function isn't allowed to return anything
 	       but inches */
 	    if(state->unit == MM) {
-		(double)curr_net->start_x /= 25.4;
-		(double)curr_net->start_y /= 25.4;
+		curr_net->start_x /= 25.4;
+		curr_net->start_y /= 25.4;
 		/* KLUDGE. All images, regardless of input format,
 		   are returned in INCH format */
 		curr_net->unit = INCH;
